requirements.txt
# global
# todo: pin to version numbers
django<1.9
django-jinja
django-pipeline
django-bower
django-markupfield
pyreact
path.py
django-toolbelt
uwsgi
jsonfield==1.0.3
markdown
django-suit

# test
tox
pytest
pytest-django

# dev
isort
autopep8
django-debug-toolbar
honcho
